# services/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hospice EHR API - Certifications & Medications
  description: |
    API for managing Medicare certifications, Face-to-Face encounters, physician orders,
    medications, and medication administration in a hospice EHR system.

    ## Key Features
    - **Medicare Compliance**: Track certification periods (INITIAL_90, SUBSEQUENT_90, SUBSEQUENT_60)
    - **Electronic Signatures**: 21 CFR Part 11 compliant with SHA-256 hashing
    - **Face-to-Face Tracking**: Required encounters within 30 days of recertification
    - **Controlled Substances**: DEA-compliant tracking and logging
    - **MAR System**: Complete medication administration documentation
    - **RBAC**: Role-based access control on all endpoints
    - **Audit Logging**: Full audit trail for HIPAA compliance

    ## Authentication
    All endpoints require Bearer token authentication with appropriate RBAC permissions.

  version: 1.0.0
  contact:
    name: API Support
    email: api-support@hospice-ehr.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://staging-api.hospice-ehr.com/api
    description: Staging server
  - url: https://api.hospice-ehr.com/api
    description: Production server
security:
  - BearerAuth: []
...
paths:
...
  /patients/{patientId}/medications:
    get:
      tags:
        - Medications
      summary: Get patient medications
      description: Retrieve all medications for a patient with optional filtering
      operationId: getPatientMedications
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/MedicationStatus'
        - name: is_hospice_related
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Medication'
                  count:
                    type: integer

    post:
      tags:
        - Medications
      summary: Create medication
      description: Create a new medication for a patient
      operationId: createMedication
      parameters:
        - $ref: '#/components/parameters/PatientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationCreate'
      responses:
        '201':
          description: Medication created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationResponse'
  /patients/{patientId}/medications/{medicationId}/discontinue:
    post:
      tags:
        - Medications
      summary: Discontinue medication
      description: Discontinue a medication with reason
      operationId: discontinueMedication
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - name: medicationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  example: 'No longer needed for symptom management'
      responses:
        '200':
          description: Medication discontinued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Medication'
  /patients/{patientId}/medications/{medicationId}/hold:
    post:
      tags:
        - Medications
      summary: Hold medication
      description: Temporarily hold a medication
      operationId: holdMedication
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - name: medicationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  example: 'Patient nausea, hold until resolved'
                hold_until:
                  type: string
                  format: date
                  example: '2024-01-25'
      responses:
        '200':
          description: Medication held
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Medication'

  # ============================================================================
  # MAR (Medication Administration Record)
  # ============================================================================
  /patients/{patientId}/mar:
    get:
      tags:
        - MAR
      summary: Get patient MAR
      description: Retrieve MAR entries for a patient with optional filtering
      operationId: getPatientMAR
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: medication_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MAREntry'
                  count:
                    type: integer

    post:
      tags:
        - MAR
      summary: Record MAR entry
      description: Record a medication administration
      operationId: createMAREntry
      parameters:
        - $ref: '#/components/parameters/PatientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MARCreate'
            examples:
              given:
                summary: Medication given
                value:
                  medication_id: 1
                  scheduled_time: '2024-01-15T10:00:00Z'
                  actual_time: '2024-01-15T10:05:00Z'
                  mar_status: GIVEN
                  dosage_given: '15mg'
                  route_used: ORAL
                  administered_by_name: 'Nurse Johnson'
                  patient_response: 'Patient tolerated medication well. Pain reduced from 8/10 to 4/10 after 30 minutes.'
              refused:
                summary: Medication refused
                value:
                  medication_id: 1
                  scheduled_time: '2024-01-15T14:00:00Z'
                  mar_status: REFUSED
                  administered_by_name: 'Nurse Johnson'
                  reason_not_given: 'Patient stated pain level only 2/10, declined medication at this time.'
      responses:
        '201':
          description: MAR entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MARResponse'

  # ============================================================================
  # COMFORT KITS
  # ============================================================================
  /patients/{patientId}/comfort-kit:
    get:
      tags:
        - Comfort Kits
      summary: Get patient comfort kits
      description: Retrieve all comfort kits for a patient
      operationId: getPatientComfortKit
      parameters:
        - $ref: '#/components/parameters/PatientId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ComfortKit'
                  count:
                    type: integer

    post:
      tags:
        - Comfort Kits
      summary: Create comfort kit
      description: Create an emergency medication kit for a patient
      operationId: createComfortKit
      parameters:
        - $ref: '#/components/parameters/PatientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComfortKitCreate'
      responses:
        '201':
          description: Comfort kit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComfortKitResponse'
  /patients/{patientId}/comfort-kit/destroy:
    post:
      tags:
        - Comfort Kits
      summary: Destroy comfort kit
      description: Destroy a comfort kit with witness logging (DEA requirement)
      operationId: destroyComfortKit
      parameters:
        - $ref: '#/components/parameters/PatientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - kit_id
              properties:
                kit_id:
                  type: integer
                witness_id:
                  type: string
                witness_name:
                  type: string
                  example: 'Pharmacist Brown'
                destruction_notes:
                  type: string
                  example: 'Patient deceased. All remaining medications destroyed per policy.'
      responses:
        '200':
          description: Comfort kit destroyed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/ComfortKit'

  # ============================================================================
  # MEDICATION RECONCILIATION
  # ============================================================================
  /patients/{patientId}/medication-reconciliation:
    post:
      tags:
        - Medication Reconciliation
      summary: Create medication reconciliation
      description: Create a CMS-required medication reconciliation
      operationId: createMedicationReconciliation
      parameters:
        - $ref: '#/components/parameters/PatientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationReconciliationCreate'
      responses:
        '201':
          description: Medication reconciliation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationReconciliationResponse'

# ============================================================================
# COMPONENTS
# ============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
  parameters:
    PatientId:
      name: patientId
      in: path
      required: true
      schema:
        type: integer
      description: Patient ID
...
  schemas:
...
    MedicationStatus:
      type: string
      enum:
        - ACTIVE
        - DISCONTINUED
        - COMPLETED
        - HELD
        - ON_HOLD
...
    Medication:
      type: object
      properties:
        id:
          type: integer
        patient_id:
          type: integer
        medication_name:
          type: string
        generic_name:
          type: string
        medication_status:
          $ref: '#/components/schemas/MedicationStatus'
        medication_route:
          type: string
          example: 'ORAL'
        dosage:
          type: string
        frequency:
          type: string
        instructions:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          nullable: true
        discontinued_date:
          type: string
          format: date
          nullable: true
        discontinuation_reason:
          type: string
          nullable: true
        controlled_schedule:
          $ref: '#/components/schemas/ControlledSchedule'
        is_hospice_related:
          type: boolean
        prescriber_id:
          type: string
        order_id:
          type: integer
        created_by_id:
          type: string
        updated_by_id:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MedicationCreate:
      type: object
      required:
        - medication_name
        - start_date
      properties:
        medication_name:
          type: string
        generic_name:
          type: string
        medication_route:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        instructions:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        controlled_schedule:
          $ref: '#/components/schemas/ControlledSchedule'
        is_hospice_related:
          type: boolean
          default: true
        prescriber_id:
          type: string
        order_id:
          type: integer
        initial_quantity:
          type: string
    MedicationResponse:
      type: object
      properties:
        status:
          type: integer
          example: 201
        message:
          type: string
          example: 'Medication created'
        data:
          $ref: '#/components/schemas/Medication'

    # ========================================================================
    # MAR
    # ========================================================================
...
    MAREntry:
      type: object
      properties:
        mar_entry:
          type: object
          properties:
            id:
              type: integer
            patient_id:
              type: integer
            medication_id:
              type: integer
            scheduled_time:
              type: string
              format: date-time
            actual_time:
              type: string
              format: date-time
            mar_status:
              $ref: '#/components/schemas/MARStatus'
            dosage_given:
              type: string
            route_used:
              type: string
            administered_by_id:
              type: string
            administered_by_name:
              type: string
            reason_not_given:
              type: string
              nullable: true
            patient_response:
              type: string
              nullable: true
            createdAt:
              type: string
              format: date-time
        medication:
          type: object
          properties:
            id:
              type: integer
            medication_name:
              type: string
            dosage:
              type: string
            frequency:
              type: string
    MARCreate:
      type: object
      required:
        - medication_id
        - scheduled_time
        - mar_status
      properties:
        medication_id:
          type: integer
        scheduled_time:
          type: string
          format: date-time
        actual_time:
          type: string
          format: date-time
        mar_status:
          $ref: '#/components/schemas/MARStatus'
        dosage_given:
          type: string
        route_used:
          type: string
        administered_by_name:
          type: string
        reason_not_given:
          type: string
          description: 'Required when mar_status is NOT_GIVEN, REFUSED, or HELD'
        patient_response:
          type: string
    MARResponse:
      type: object
      properties:
        status:
          type: integer
          example: 201
        message:
          type: string
          example: 'MAR entry created'
        data:
          type: object

    # ========================================================================
    # COMFORT KITS
    # ========================================================================
    ComfortKit:
      type: object
      properties:
        id:
          type: integer
        patient_id:
          type: integer
        kit_number:
          type: string
        issue_date:
          type: string
          format: date
        expiration_date:
          type: string
          format: date
        status:
          type: string
          enum: [ACTIVE, EXPIRED, DESTROYED, RETURNED]
        medications:
          type: array
          items:
            type: object
            properties:
              medication:
                type: string
              quantity:
                type: string
              lot_number:
                type: string
              ndc:
                type: string
        location:
          type: string
        created_by_id:
          type: string
        updated_by_id:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ComfortKitResponse:
      type: object
      properties:
        status:
          type: integer
          example: 201
        message:
          type: string
          example: 'Comfort kit created'
        data:
          $ref: '#/components/schemas/ComfortKit'

    # ========================================================================
    # MEDICATION RECONCILIATION
    # ========================================================================
...